feat: init the content
Prior to this commit, the page is still the default layout of Next.js.
I will start updating the ui from this commit.

In this commit, I updated the title and meta description, and
temporarily filled the web page with some texts generated by chatGPT,
this will be updated in future commits.