feat: build static ui (#5)
Prior to this commit, the page is still the default layout of Next.js.
I will start updating the ui from this commit.

In this commit, I 
1. update the title and meta description, 

2. add dependencies of `"@emotion/react": "11.10.6"`,
`"@emotion/styled": "11.10.6"`, and
`"@mui/material": "5.12.0"`.

3. upload a new png picture for temporarily profile picture usage,
and remove the redundant svg files as well.

4. add `AboutMeSection`, `ExperienceSection`, `EducationSection`,
`ProjectSection` and `SkillSection`
They are not yet implemented and styled well. I will update them
step by step in future commits.