build(eslint): add react eslint plugin
Add `eslint-plugin-react` for this project. This commit also fix those
coming eslint error:

1. `react/jsx-key`

Error: Missing "key" prop for element in iterator

We have to add `key` prop for array in react.

2. `react/no-unescaped-entities`

I have some `'` in used which should be replaced by `&apos;`.

This commit also ban one of the rules, which is unnecessary for next.js
project, that is `react/react-in-jsx-scope`[1].

[1]: https://stackoverflow.com/a/61160875